% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{linesInRemoteFile}
\alias{linesInRemoteFile}
\title{Return the number of lines in a file downloaded to a temp dir.}
\usage{
linesInRemoteFile(url, dir)
}
\arguments{
\item{url}{...}

\item{dir}{...}
}
\value{
return the number of lines in the file
}
\description{
Download the file to a local directory, returning the number of
lines in that file.  This is a toy function, to shed light on
https://github.com/r-lib/pkgdown/issues/2774.
}
\examples{
dir <- tempdir()
url <- "https://raw.githubusercontent.com/dankelley/dod/main/NEWS.md"
linesInRemoteFile(url, dir)
unlink(dir, recursive = TRUE)

}
